remove unnecessary datastore interface
