+++ cmd/api.go
@@ -75,7 +75,7 @@ var apiCmd = &cobra.Command{
 			log.WithError(err).Fatalf("Failed to connect to Postgres database at %s", postgresDSN)
 		}
 
-		ds, err := datastore.NewProdDatastore(log, redis, db)
+		ds, err := datastore.NewDatastore(log, redis, db)
 		if err != nil {
 			log.WithError(err).Fatalf("Failed setting up prod datastore")
 		}
@@ -96,6 +96,7 @@ var apiCmd = &cobra.Command{
 			BeaconClient:  beaconClient,
 			Datastore:     ds,
 			Redis:         redis,
+			DB:            db,
 			EthNetDetails: *networkInfo,
 			PprofAPI:      apiPprofEnabled,
 			SecretKey:     sk,
